Make UserAddressForm get AbstractAddressForm (subclass) dynamically.
