+++ oscar/apps/address/abstract_forms.py
@@ -1,0 +1,14 @@
+from django.conf import settings
+from django import forms
+
+class AbstractAddressForm(forms.ModelForm):
+
+    def __init__(self, *args, **kwargs):
+        """
+        Set fields in OSCAR_REQUIRED_ADDRESS_FIELDS as required.
+        """
+        super(AbstractAddressForm, self).__init__(*args, **kwargs)
+        field_names = (set(self.fields) &
+                       set(settings.OSCAR_REQUIRED_ADDRESS_FIELDS))
+        for field_name in field_names:
+            self.fields[field_name].required = True
