+++ oscar/apps/address/forms.py
@@ -1,23 +1,10 @@
-from django.conf import settings
-from django import forms
 
-from oscar.core.loading import get_model
+from oscar.core.loading import get_class, get_model
 from oscar.views.generic import PhoneNumberMixin
 
 UserAddress = get_model('address', 'useraddress')
 
-
-class AbstractAddressForm(forms.ModelForm):
-
-    def __init__(self, *args, **kwargs):
-        """
-        Set fields in OSCAR_REQUIRED_ADDRESS_FIELDS as required.
-        """
-        super(AbstractAddressForm, self).__init__(*args, **kwargs)
-        field_names = (set(self.fields) &
-                       set(settings.OSCAR_REQUIRED_ADDRESS_FIELDS))
-        for field_name in field_names:
-            self.fields[field_name].required = True
+AbstractAddressForm = get_class('address.abstract_forms', 'AbstractAddressForm')
 
 
 class UserAddressForm(PhoneNumberMixin, AbstractAddressForm):
